% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_capsules.R
\name{get_SpaceX_capsules}
\alias{get_SpaceX_capsules}
\title{GET info about SpaceX capsules}
\usage{
get_SpaceX_capsules(name = NULL)
}
\arguments{
\item{name}{\link{character}: Name of the capsule. Choose between "dragon1", "dragon2" or "crewdragon".
By default all three rocket data are returned.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX capsules
}
