% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_cores.R
\name{get_SpaceX_cores}
\alias{get_SpaceX_cores}
\title{GET info about SpaceX cores}
\usage{
get_SpaceX_cores(type = NULL, ...)
}
\arguments{
\item{type}{\link{character}: type of the request: Choose between "upcoming" or "past".}

\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX cores
}
\examples{

# get all cores
get_SpaceX_cores()

# get core by serial number
get_SpaceX_cores(core_serial = "B1037")


}
