% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_dragons.R
\name{get_SpaceX_dragons}
\alias{get_SpaceX_dragons}
\title{GET info about SpaceX all dragon capsules}
\usage{
get_SpaceX_dragons(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX all dragon capsules
}
\examples{

# get all data for dragon capsules
get_SpaceX_dragons()

# get data for "dragon1"
get_SpaceX_dragons(id = "dragon1")

}
