% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_history.R
\name{get_SpaceX_history}
\alias{get_SpaceX_history}
\title{GET info about SpaceX history}
\usage{
get_SpaceX_history(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX history
}
\examples{

# get all data
get_SpaceX_history()

get_SpaceX_history(id = 17)

# get data with flight_number 41
get_SpaceX_history(flight_number = 41)

}
