% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_landpads.R
\name{get_SpaceX_landpads}
\alias{get_SpaceX_landpads}
\title{GET info about SpaceX landpads}
\usage{
get_SpaceX_landpads(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX landpads
}
\examples{

# get all all landpads
get_SpaceX_landpads()

# get landing pads by id
get_SpaceX_landpads(id = "LZ-4")

}
