% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_parts.R
\name{get_SpaceX_parts}
\alias{get_SpaceX_parts}
\title{GET detailed info about SpaceX parts}
\usage{
get_SpaceX_parts(type, name = NULL, ...)
}
\arguments{
\item{type}{\link{character}: Type of parts. Choose between "caps" for capsule details
or "cores" for core details.}

\item{name}{\link{character}: Name of capsule or core.}

\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET detailed info about SpaceX parts
}
\details{
Query Strings for \code{type = "caps"}
\tabular{rr}{
Query Strings \tab	Description  \cr
capsule_serial \tab	Filter by capsule serial # \cr
capsule_id \tab	Filter by capsule id \cr
status \tab	Filter by capsule status \cr
original_launch \tab	Filter by original launch date \cr
missions \tab	Filter by flight missions \cr
landings \tab	Filter by # of landings \cr
type \tab	Filter by type of dragon capusle \cr
}

Query Strings for \code{type = "cores"}
\tabular{rr}{
core_serial \tab	Filter by core serial # \cr
status \tab	Filter by flight status \cr
original_launch \tab	Filter by original launch date \cr
missions \tab	Filter by flight missions \cr
rtls_attempt \tab	Filter by at least 1 rtls attempt \cr
rtls_landings \tab	Filter by total number of landings \cr
asds_attempt \tab	Filter by at least 1 asds attempt \cr
asds_landings \tab	Filter by total number of landings \cr
water_landing \tab	Filter by at least 1 water landing \cr
}
}
\examples{

## get details about all cores
get_SpaceX_parts(type = "cores")

## get details about all capsules
get_SpaceX_parts(type = "caps")

## get details about all active capsules
get_SpaceX_parts(type = "caps", status = "active")

## get details about core B1041
get_SpaceX_parts(type = "cores", name = "B1041")

}
