% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_payloads.R
\name{get_SpaceX_payloads}
\alias{get_SpaceX_payloads}
\title{GET info about SpaceX payloads}
\usage{
get_SpaceX_payloads(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX payloads
}
\examples{

get_SpaceX_payloads()

get_SpaceX_payloads(payload_id = "Telkom-4")

}
