% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpcaeX_roadster.R
\name{get_SpaceX_roadster}
\alias{get_SpaceX_roadster}
\title{GET info about gives orbital parameters about the current position of Elon's Tesla Roadster
that was launched on the first Falcon Heavy demo flight.}
\usage{
get_SpaceX_roadster(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about gives orbital parameters about the current position of Elon's Tesla Roadster
that was launched on the first Falcon Heavy demo flight.
}
\examples{

## get details about the Tesla roadster position
get_SpaceX_roadster()

}
