% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_rockets.R
\name{get_SpaceX_rockets}
\alias{get_SpaceX_rockets}
\title{GET info about SpaceX rockets}
\usage{
get_SpaceX_rockets(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX rockets
}
\examples{

## get details about all rockets
get_SpaceX_rockets()

## get details about "falcon1" rocket
get_SpaceX_rockets(rocket_id = "falcon1")

}
