% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_ships.R
\name{get_SpaceX_ships}
\alias{get_SpaceX_ships}
\title{GET info about SpaceX missions}
\usage{
get_SpaceX_ships(...)
}
\arguments{
\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX missions
}
\examples{

# get all SpaceX ships
get_SpaceX_ships()

get_SpaceX_ships(ship_id = "AMERICANCHAMPION")

}
