% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqlseq_modified.R
\name{pqlseq_modified}
\alias{pqlseq_modified}
\title{TITLE}
\usage{
pqlseq_modified(
  RawCountDataSet,
  Phenotypes,
  Covariates = NULL,
  RelatednessMatrix = NULL,
  LibSize = NULL,
  fit.model = "PMM",
  fit.method = "AI.REML",
  fit.maxiter = 500,
  fit.tol = 1e-05,
  numCore = 1,
  filtering = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{RawCountDataSet}{DESCRIPTION}

\item{Phenotypes}{DESCRIPTION}

\item{Covariates}{DESCRIPTION}

\item{RelatednessMatrix}{DESCRIPTION}

\item{LibSize}{DESCRIPTION}

\item{fit.model}{DESCRIPTION}

\item{fit.method}{DESCRIPTION}

\item{fit.maxiter}{DESCRIPTION}

\item{fit.tol}{DESCRIPTION}

\item{numCore}{DESCRIPTION}

\item{filtering}{DESCRIPTION}

\item{verbose}{DESCRIPTION}

\item{...}{DESCRIPTION}
}
\value{
A LIST
}
\description{
DESCRIPTION
}
