% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceX.R
\docType{package}
\name{SpaceX-package}
\alias{SpaceX-package}
\title{SpaceX}
\description{
An R API wrapper for the SpaceX API.
}
\details{
\tabular{ll}{Package: \tab SpaceX\cr Type: \tab Package\cr Version:
\tab 0.0.1 \cr Date: \tab 2020-XX-XX \cr License: \tab GPL-2\cr }
}
\keyword{package}
