% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_launches.R
\name{get_SpaceX_launches}
\alias{get_SpaceX_launches}
\title{GET info about SpaceX launches}
\usage{
get_SpaceX_launches(type = NULL, ...)
}
\arguments{
\item{type}{\link{character}: type of the request: Choose between "upcoming", "past", "latest" or "next".}

\item{...}{: Query string to API. See more in details.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX launches
}
\details{

}
\examples{

## get details about all past launches
get_SpaceX_launches()

## get details about all upcoming launches
get_SpaceX_launches(type = "upcoming")

## get details about the lastes launch
get_SpaceX_launches(type = "latest")

## get details about all past launches in 2018
get_SpaceX_launches(launch_year = 2018)

## get details about a special flight_id
get_SpaceX_launches(flight_id = "5a9fc479ab70786ba5a1eaaa")

}
