% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SpaceX_rockets.R
\name{get_SpaceX_rockets}
\alias{get_SpaceX_rockets}
\title{GET info about SpaceX rockets}
\usage{
get_SpaceX_rockets(rocket_id = NULL, ...)
}
\arguments{
\item{name}{\link{character}: Name of the rocket. Choose between "falcon1", "falcon9","falconheavy" or "starship".
By default all three rocket data are returned.}
}
\value{
\link{data.frame} with parsed data from SpaceX API.
}
\description{
GET info about SpaceX rockets
}
\examples{

## get details about all rockets
get_SpaceX_rockets()

## get details about "falcon1" rocket
get_SpaceX_rockets("falcon1")

}
