% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceX.R
\name{SpaceX}
\alias{SpaceX}
\title{Estimation of shared and cluster specific gene co-expression networks for spatial transcriptomics data.}
\usage{
SpaceX(
  Gene_expression_mat,
  Spatial_locations,
  Cluster_annotations,
  sPMM = FALSE,
  Post_process = FALSE,
  numCore = 1,
  nrun = 10000,
  burn = 5000
)
}
\arguments{
\item{Gene_expression_mat}{Gene expression dataframe (N X G).}

\item{Spatial_locations}{Spatial locations with coordinates. This should be provided as dataframe.}

\item{Cluster_annotations}{Cluster annotations for each of the spatial location.}

\item{sPMM}{If \code{TRUE}, the code will return the estimates of sigma1_sq and sigma2_sq from the spatial Poisson mixed model.}

\item{Post_process}{If \code{FALSE}, the code will return the posterior samples of \code{Phi} and \code{Psi^c} (based on definition in equation 1 of the SpaceX paper) only.
Default is \code{TRUE} and the code will return all the posterior samples, shared and cluster specific co-expressions.}

\item{numCore}{The number of cores for parallel computing (default = 1).}

\item{nrun}{The total number of MCMC iterations for MSFA. Default is 10000.}

\item{burn}{The number of burnin samples. Default is 5000.}
}
\value{
\item{Posterior_samples}{Posterior samples}
\item{Shared_network}{Shared co-expression matrix}
\item{Cluster_network}{Cluster specific co-expression matrices}
}
\description{
SpaceX function estimates shared and cluster specific gene co-expression networks for spatial transcriptomics data. Please make sure to provide both inputs as dataframe. More details about the SpaceX algorithm can be found in the reference paper.
}
\examples{
Implementation details and examples can be found at this link https://bookdown.org/satwik91/SpaceX_supplementary/.


}
\references{
Acharyya S., Zhou X., Baladandayuthapani V. (2021). SpaceX: Gene Co-expression Network Estimation for Spatial Transcriptomics.
}
