% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceX.R
\name{SpaceX}
\alias{SpaceX}
\title{SpaceX function}
\usage{
SpaceX(Gene_expression_mat, Spatial_locations)
}
\arguments{
\item{Gene_expression_mat}{Gene expression matrix (N X G) [Should be provided as dataframe]}

\item{Spatial_locations}{Spatial locations and cluster annotations. [Should be provided as dataframe]
(x,y) coordinates on the first two columns and
cluster annotations on the third column.}
}
\value{
A LIST
}
\description{
DESCRIPTION
}
